MB-69799: Skip redundant DNS query during IPv6 retry
If the initial SMTP connect fails with nxdomain error, we should not
repeat the DNS queries. The DNS query is unnecessary as the set of
Hosts does not change between retries. This commit preserves the Hosts
we've retrieved and retries just the socket connect with [inet6]
option.

Previously, we incurred the DNS latency twice every time we retried an
email alert with IPv6 socket option. With slow/unresponsive DNS name
servers, this can push us over the 15-second timeout in ns_mail.